Restructure config/tools in utils
